Refactor into smaller components
